Split feature-gated library support into separate crates
Move serde/fake/sqlx support from kubetsu feature flags into independent
crates (kubetsu-serde, kubetsu-fake, kubetsu-sqlx) so that each adapter
can release and version independently from the core crate.

- kubetsu: core crate with define_id! macro and KubetsuId marker trait
- kubetsu-serde: impl_serde! macro for serde Serialize/Deserialize
- kubetsu-fake: impl_fake! macro for fake::Dummy
- kubetsu-sqlx: impl_sqlx! macro with any/mysql/postgres/sqlite features
- kubetsu-tests: integration tests combining all adapter crates
- Fix LICENSE filename typo (LISENCE -> LICENSE)
- Add UPGRADE.md with migration guide from 0.6.x to 0.7.0
