+++ Cargo.toml
@@ -1,41 +1,9 @@
-[package]
-name = "kubetsu"
-version = "0.6.0"
-edition = "2021"
-
-authors = ["Keiji Yoshimi"]
-
-readme = "README.md"
-
-license = "MIT"
-
-description = "distinguish value type of other struct"
-
-homepage = "https://github.com/walf443/kubetsu"
-repository = "https://github.com/walf443/kubetsu.git"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[features]
-default  = []
-all = ["serde", "fake", "sqlx-mysql", "sqlx-sqlite", "sqlx-postgres", "sqlx-any"]
-serde = ["dep:serde"]
-fake = ["dep:fake"]
-sqlx = ["dep:sqlx"]
-sqlx-any = ["sqlx", "sqlx/any"]
-sqlx-mysql = ["sqlx", "sqlx/mysql"]
-sqlx-postgres = ["sqlx", "sqlx/postgres"]
-sqlx-sqlite = ["sqlx", "sqlx/sqlite"]
-
-[dependencies]
-serde = { version = "1", default-features = false, optional = true }
-fake  = { version = "5", features = ["derive"], optional = true }
-sqlx  = { version = "0.8", default-features = false, features = [], optional = true }
-
-[dev-dependencies]
-sqlx  = { version = "0.8", features = ["runtime-tokio", "macros", "sqlite"] }
-tokio  = { version = "1", features = ["macros", "rt"] }
-serde_json = { version = "1" }
-testcontainers = { version = "0.27" }
-testcontainers-modules = { version = "0.15", features = ["postgres", "mysql"] }
-ctor = { version = "0.8" }
+[workspace]
+resolver = "2"
+members = [
+    "kubetsu",
+    "kubetsu-serde",
+    "kubetsu-fake",
+    "kubetsu-sqlx",
+    "kubetsu-tests",
+]
