+++ kubetsu-serde/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "kubetsu-serde"
+version = "0.1.0"
+edition = "2021"
+
+authors = ["Keiji Yoshimi"]
+
+license = "MIT"
+
+description = "serde support for kubetsu ID types"
+
+homepage = "https://github.com/walf443/kubetsu"
+repository = "https://github.com/walf443/kubetsu.git"
+
+[dependencies]
+kubetsu = { version = "0.7", path = "../kubetsu" }
+serde = { version = "1", default-features = false }
+
+[dev-dependencies]
+kubetsu = { path = "../kubetsu" }
+serde_json = { version = "1" }
