+++ kubetsu-sqlx/Cargo.toml
@@ -1,0 +1,29 @@
+[package]
+name = "kubetsu-sqlx"
+version = "0.1.0"
+edition = "2021"
+
+authors = ["Keiji Yoshimi"]
+
+license = "MIT"
+
+description = "sqlx support for kubetsu ID types"
+
+homepage = "https://github.com/walf443/kubetsu"
+repository = "https://github.com/walf443/kubetsu.git"
+
+[features]
+default = []
+any = ["sqlx/any"]
+mysql = ["sqlx/mysql"]
+postgres = ["sqlx/postgres"]
+sqlite = ["sqlx/sqlite"]
+
+[dependencies]
+kubetsu = { version = "0.7", path = "../kubetsu" }
+sqlx = { version = "0.8", default-features = false, features = [] }
+
+[dev-dependencies]
+kubetsu = { path = "../kubetsu" }
+sqlx = { version = "0.8", features = ["runtime-tokio", "macros", "sqlite"] }
+tokio = { version = "1", features = ["macros", "rt"] }
