+++ kubetsu-tests/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "kubetsu-tests"
+version = "0.0.0"
+edition = "2021"
+publish = false
+
+[dependencies]
+kubetsu = { path = "../kubetsu" }
+kubetsu-serde = { path = "../kubetsu-serde" }
+kubetsu-fake = { path = "../kubetsu-fake" }
+kubetsu-sqlx = { path = "../kubetsu-sqlx", features = ["sqlite"] }
+serde_json = "1"
+fake = { version = "5", features = ["derive"] }
+sqlx = { version = "0.8", features = ["runtime-tokio", "macros", "sqlite"] }
+tokio = { version = "1", features = ["macros", "rt"] }
