+++ kubetsu/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "kubetsu"
+version = "0.7.0"
+edition = "2021"
+
+authors = ["Keiji Yoshimi"]
+
+readme = "README.md"
+
+license = "MIT"
+
+description = "distinguish value type of other struct"
+
+homepage = "https://github.com/walf443/kubetsu"
+repository = "https://github.com/walf443/kubetsu.git"
+
+include = ["src/**/*", "Cargo.toml", "README.md", "UPGRADE.md", "LICENSE"]
