+++ kubetsu/src/id/test.rs
@@ -1,0 +1,3 @@
+mod float;
+mod int;
+mod str;
