+++ kubetsu/src/id/test/float.rs
@@ -1,0 +1,47 @@
+use crate::Id;
+
+struct Foo {}
+
+#[test]
+fn test_partial_eq_f64() {
+    let id1: Id<Foo, f64> = Id::new(1.0);
+    let id2: Id<Foo, f64> = Id::new(1.0);
+    let id3: Id<Foo, f64> = Id::new(2.0);
+
+    assert_eq!(id1, id2);
+    assert_ne!(id1, id3);
+}
+
+#[test]
+fn test_partial_eq_f32() {
+    let id1: Id<Foo, f32> = Id::new(1.0);
+    let id2: Id<Foo, f32> = Id::new(1.0);
+    let id3: Id<Foo, f32> = Id::new(2.0);
+
+    assert_eq!(id1, id2);
+    assert_ne!(id1, id3);
+}
+
+#[test]
+fn test_inner_f64() {
+    let id1: Id<Foo, f64> = Id::new(1.0);
+    assert_eq!(id1.inner(), &1.0);
+}
+
+#[test]
+fn test_inner_f32() {
+    let id1: Id<Foo, f32> = Id::new(1.0);
+    assert_eq!(id1.inner(), &1.0);
+}
+
+#[test]
+fn test_clone_f64() {
+    let id1: Id<Foo, f64> = Id::new(1.0);
+    assert_eq!(id1.clone(), id1);
+}
+
+#[test]
+fn test_clone_f32() {
+    let id1: Id<Foo, f32> = Id::new(1.0);
+    assert_eq!(id1.clone(), id1);
+}
