+++ kubetsu/src/lib.rs
@@ -1,0 +1,19 @@
+mod macros;
+
+mod id;
+#[doc = include_str!("../README.md")]
+#[deprecated(
+    since = "0.6.0",
+    note = "Use `kubetsu::define_id!` to define your own ID type instead."
+)]
+pub use id::Id;
+
+/// Marker trait for types generated by [`define_id!`].
+///
+/// Adapter crates (`kubetsu-serde`, `kubetsu-fake`, `kubetsu-sqlx`) require
+/// this trait to ensure that their macros are only used with kubetsu ID types.
+pub trait KubetsuId {
+    type Inner;
+    fn new(inner: Self::Inner) -> Self;
+    fn inner(&self) -> &Self::Inner;
+}
