--- src/id/test.rs
@@ -1,5 +1,0 @@
-mod float;
-mod int;
-#[cfg(feature = "sqlx")]
-mod sqlx;
-mod str;
