--- src/id/test/sqlx.rs
@@ -1,11 +1,0 @@
-#[cfg(feature = "sqlx-mysql")]
-mod mysql;
-
-#[cfg(feature = "sqlx-postgres")]
-mod postgres;
-
-#[cfg(feature = "sqlx-sqlite")]
-mod sqlite;
-
-#[cfg(feature = "sqlx-any")]
-mod any;
